requirements.txt
supervisely==6.2.8
trimesh
pynrrd